requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==22.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
albucore==0.0.13
albumentations==1.4.13
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0

blinker==1.8.2

cachetools==5.3.3
certifi==2017.4.17
cffi==1.16.0
chainmap==1.0.3
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
cmake==3.30.2
combomethod==1.0.12
comm==0.2.2
conda-pack==0.8.0
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0

decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1

Flask==3.0.3
flow-vis==0.1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43

google-ai-generativelanguage==0.1.0
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.65.0
GPUtil==1.4.0
-e git+https://github.com/IDEA-Research/GroundingDINO.git@e27a646ca0ea29c7883187c6f7303cc975366679#egg=groundingdino
grpcio==1.66.1
grpcio-status==1.62.3
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.0
hydra-core==1.1.0

ifcfg==0.21
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.3
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1

jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9

jupyter_packaging==0.12.3
jupyter_server==2.14.0
jupyter_server_fileid==0.9.1
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
lit==18.1.8
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.7.2
Markdown==3.6

matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2

mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.4
nulltype==2.3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.0.0
objaverse==0.1.7
omegaconf==2.1.2
open3d==0.18.0
openai==1.45.1
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10

plotly==5.22.0


progressbar==2.5
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycosat==0.6.6
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyDeprecate==0.3.2
pyglet==1.5.29
Pygments==2.17.2
pyparsing==3.1.2
pyquaternion==0.9.9
pyrallis==0.3.1

python-json-logger==2.0.7
pytorch-lightning==0.7.1
pytorch3d==0.7.5
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==1.45.0
setproctitle==1.3.3

smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
supervision==0.21.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.7
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.4


transformations==2022.9.26
transformers==4.42.4
trimesh==4.4.1
triton==2.0.0
ttach==0.0.3
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0

wandb==0.16.6

webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.13
y-py==0.6.2
yapf==0.40.2
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.19.0
